Don't free the stacklet handle on threadlet dealloc
The threadlet will remain alive until it's done, in which case it will
have been reset to EMPTY_STACKLET_HANDLE anyway. If a threadlet is not
switched to anymore, it would leak the stacklet handle, this doesn't
matter if it happens on interpreter shutdown, but it may happen when a
thread goes away, so we need a way to fix this.