Clear frame on dealloc
It's not 100% clear to me if this is kosher, but lets see how it goes.
If a fiber is deallocated and it did have an active frame, I guess it's
ok to clear it, since it's unreachable (it could be part of a frame
chain though), and this could help by relesing variables local to that
frame and have them gc'd.